Fix all doc prefixes
Currently, there are a lot of 404s in the docs because of incorrect path
prefixes. Most of the links assume that the docs are served at the root
of the docs site, but actually there's an `/ome` prefix.

Also, in places where docs don't actually exist, I've created a place
holder `Coming soon!` instead of users confusingly getting a 404.