Add LWS worker-index 0 to sample DS manifest
If you don't set `leaderworkerset.sigs.k8s.io/worker-index=0` for the
router selector like this, then the behavior of the router will
potentially match against one of the non-head pods which is invalid. The
non-head pod will return a 404 and your generations will stall.

Actually, other example manifests have this already. It was just this
one that seems to be missing this label expression.